requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.110.2
fastdup==1.120
fuzzywuzzy==0.18.0
h11==0.14.0
idna==3.7
importlib-metadata==7.1.0
ipykernel==6.29.4
ipython==8.18.1
jedi==0.19.1
jupyter-client==8.6.1
jupyter-core==5.7.2
Levenshtein==0.25.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
pillow-heif==0.16.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.1
pydantic-core==2.18.2
pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.2
rapidfuzz==3.8.1
requests==2.31.0
sentry-sdk==2.0.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
wcwidth==0.2.13
zipp==3.18.1
